move access token to constructor remove username in challange
